fix: integration pages permissions on load children module
