+++ apps/gauzy/src/app/@shared/components/index.ts
@@ -1,4 +1,3 @@
 import { AvatarComponent } from './avatar/avatar.component';
-import { WorkspaceComponent } from './workspace/workspace.component';
 
-export const Components = [AvatarComponent, WorkspaceComponent];
+export const Components = [AvatarComponent];
