--- apps/gauzy/src/app/@shared/components/workspace/workspace.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'ngx-integrations-workspace',
-	template: '<router-outlet></router-outlet>'
-})
-export class WorkspaceComponent implements OnInit {
-	constructor() { }
-
-	ngOnInit() { }
-}
