--- apps/gauzy/src/app/pages/integrations/components/integrations-list/integrations-list.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { IntegrationsListComponent } from './integrations-list.component';
-
-describe('IntegrationsListComponent', () => {
-	let component: IntegrationsListComponent;
-	let fixture: ComponentFixture<IntegrationsListComponent>;
-
-	beforeEach(async(() => {
-		TestBed.configureTestingModule({
-			declarations: [IntegrationsListComponent],
-			teardown: { destroyAfterEach: false }
-		}).compileComponents();
-	}));
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(IntegrationsListComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it('should create', () => {
-		expect(component).toBeTruthy();
-	});
-});
