--- apps/gauzy/src/app/pages/integrations/components/integrations/integrations.component.html
@@ -1,1 +1,0 @@
-<!--  can be used for integrated integrations -->
