--- apps/gauzy/src/app/pages/integrations/components/integrations/integrations.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'ngx-integrations',
-	templateUrl: './integrations.component.html'
-})
-export class IntegrationsComponent implements OnInit {
-	constructor() {}
-
-	ngOnInit() {}
-}
