+++ apps/gauzy/src/app/pages/integrations/integrations-routing.module.ts
@@ -1,82 +1,29 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { NgxPermissionsGuard } from 'ngx-permissions';
-import { PermissionsEnum } from '@gauzy/contracts';
-import { IntegrationsListComponent } from './components/integrations-list/integrations-list.component';
-import { WorkspaceComponent } from '../../@shared/components/workspace/workspace.component';
+import { IntegrationsComponent } from './integrations.component';
 
 const routes: Routes = [
 	{
 		path: '',
-		component: WorkspaceComponent,
-		canActivate: [NgxPermissionsGuard],
-		data: {
-			permissions: {
-				only: [PermissionsEnum.INTEGRATION_VIEW],
-				redirectTo: '/pages/dashboard'
-			}
-		},
-		children: [
-			{
-				path: '',
-				redirectTo: 'list',
-				pathMatch: 'full'
-			},
-			{
-				path: 'list',
-				component: IntegrationsListComponent,
-				data: {
-					selectors: {
-						project: false,
-						employee: false,
-						organization: false,
-						date: false
-					}
-				}
-			},
-			{
-				path: 'upwork',
-				loadChildren: () => import('../upwork/upwork.module').then(
-					(m) => m.UpworkModule
-				),
-				data: {
-					selectors: {
-						project: false,
-						employee: false,
-						organization: false,
-						date: false
-					}
-				}
-			},
-			{
-				path: 'hubstaff',
-				loadChildren: () => import('../hubstaff/hubstaff.module').then(
-					(m) => m.HubstaffModule
-				),
-				data: {
-					selectors: {
-						project: false,
-						employee: false,
-						organization: false,
-						date: false
-					}
-				}
-			},
-			{
-				path: 'gauzy-ai',
-				loadChildren: () => import('./gauzy-ai/gauzy-ai.module').then(
-					(m) => m.GauzyAIModule
-				),
-				data: {
-					selectors: {
-						project: false,
-						employee: false,
-						organization: false,
-						date: false
-					}
-				}
-			}
-		]
+		component: IntegrationsComponent
+	},
+	{
+		path: 'upwork',
+		loadChildren: () => import('../upwork/upwork.module').then(
+			(m) => m.UpworkModule
+		)
+	},
+	{
+		path: 'hubstaff',
+		loadChildren: () => import('../hubstaff/hubstaff.module').then(
+			(m) => m.HubstaffModule
+		)
+	},
+	{
+		path: 'gauzy-ai',
+		loadChildren: () => import('./gauzy-ai/gauzy-ai.module').then(
+			(m) => m.GauzyAIModule
+		)
 	}
 ];
