+++ apps/gauzy/src/app/pages/integrations/integrations.component.scss
