+++ apps/gauzy/src/app/pages/integrations/integrations.component.spec.ts
@@ -1,12 +1,11 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { IntegrationsComponent } from './integrations.component';
 
 describe('IntegrationsComponent', () => {
 	let component: IntegrationsComponent;
 	let fixture: ComponentFixture<IntegrationsComponent>;
 
-	beforeEach(async(() => {
+	beforeEach(waitForAsync(() => {
 		TestBed.configureTestingModule({
 			declarations: [IntegrationsComponent],
 			teardown: { destroyAfterEach: false }
